+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarItinerarySlice.ts
@@ -19,5 +19,8 @@ export interface CfarItinerarySlice {
      * A list of segments which make up the slice of the fare
      */
     segments: Array<CfarItinerarySliceSegment>;
+    /**
+     * Fare brand of the slice
+     */
     fareBrand?: string;
 }