+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarItinerarySliceSegment.ts
@@ -40,5 +40,8 @@ export interface CfarItinerarySliceSegment {
      */
     validatingCarrierCode: string;
     fareClass: FareClass;
+    /**
+     * Fare brand of the segment
+     */
     fareBrand?: string;
 }